+++ src/api/endpoints/tokens/get-tokens/v4.ts
@@ -108,6 +108,7 @@ export const getTokensV4Options: RouteOptions = {
             .required(),
           name: Joi.string().allow(null, ""),
           image: Joi.string().allow(null, ""),
+          media: Joi.string().allow(null, ""),
           collection: Joi.object({
             id: Joi.string().allow(null),
             name: Joi.string().allow(null, ""),
@@ -139,6 +140,7 @@ export const getTokensV4Options: RouteOptions = {
           "t"."token_id",
           "t"."name",
           "t"."image",
+          "t"."media",
           "t"."collection_id",
           "c"."name" as "collection_name",
           "t"."floor_sell_source_id",
@@ -170,22 +172,21 @@ export const getTokensV4Options: RouteOptions = {
       }
 
       if (query.attributes) {
-        const attributes: { key: string; value: string }[] = [];
-        Object.entries(query.attributes).forEach(([key, values]) => {
-          (Array.isArray(values) ? values : [values]).forEach((value) =>
-            attributes.push({ key, value })
-          );
-        });
+        const attributes: { key: string; value: any }[] = [];
+        Object.entries(query.attributes).forEach(([key, value]) => attributes.push({ key, value }));
 
         for (let i = 0; i < attributes.length; i++) {
+          const multipleSelection = Array.isArray(attributes[i].value);
+
           (query as any)[`key${i}`] = attributes[i].key;
           (query as any)[`value${i}`] = attributes[i].value;
+
           baseQuery += `
             JOIN "token_attributes" "ta${i}"
               ON "t"."contract" = "ta${i}"."contract"
               AND "t"."token_id" = "ta${i}"."token_id"
               AND "ta${i}"."key" = $/key${i}/
-              AND "ta${i}"."value" = $/value${i}/
+              AND "ta${i}"."value" ${multipleSelection ? `IN ($/value${i}:csv/)` : `= $/value${i}/`}
           `;
         }
       }
@@ -406,6 +407,7 @@ export const getTokensV4Options: RouteOptions = {
           tokenId: r.token_id,
           name: r.name,
           image: r.image,
+          media: r.media,
           collection: {
             id: r.collection_id,
             name: r.collection_name,