+++ src/models/transactions/index.ts
@@ -1,4 +1,5 @@
 import { idb } from "@/common/db";
+import { logger } from "@/common/logger";
 import { fromBuffer, toBuffer } from "@/common/utils";
 
 export type Transaction = {
@@ -9,6 +10,9 @@ export type Transaction = {
   data: string;
   blockNumber: number;
   blockTimestamp: number;
+  gasPrice?: string;
+  gasUsed?: string;
+  gasFee?: string;
 };
 
 export const saveTransaction = async (transaction: Transaction) => {
@@ -21,7 +25,10 @@ export const saveTransaction = async (transaction: Transaction) => {
         value,
         data,
         block_number,
-        block_timestamp
+        block_timestamp,
+        gas_price,
+        gas_used,
+        gas_fee
       ) VALUES (
         $/hash/,
         $/from/,
@@ -29,7 +36,10 @@ export const saveTransaction = async (transaction: Transaction) => {
         $/value/,
         $/data/,
         $/blockNumber/,
-        $/blockTimestamp/
+        $/blockTimestamp/,
+        $/gasPrice/,
+        $/gasUsed/,
+        $/gasFee/
       )
       ON CONFLICT DO NOTHING
     `,
@@ -41,6 +51,9 @@ export const saveTransaction = async (transaction: Transaction) => {
       data: toBuffer(transaction.data),
       blockNumber: transaction.blockNumber,
       blockTimestamp: transaction.blockTimestamp,
+      gasPrice: transaction.gasPrice,
+      gasUsed: transaction.gasUsed,
+      gasFee: transaction.gasFee,
     }
   );
 
@@ -62,10 +75,15 @@ export const getTransaction = async (
     { hash: toBuffer(hash) }
   );
 
-  return {
-    hash,
-    from: fromBuffer(result.from),
-    to: fromBuffer(result.to),
-    value: result.value,
-  };
+  try {
+    return {
+      hash,
+      from: fromBuffer(result.from),
+      to: fromBuffer(result.to),
+      value: result.value,
+    };
+  } catch (error) {
+    logger.info("debug", `Error fetching transaction ${hash}: ${JSON.stringify(result)}`);
+    throw error;
+  }
 };